challenge2/requirements.txt
blinker==1.7.0
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
flask==3.0.0
flask-sqlalchemy==3.1.1
fonttools==4.47.0
greenlet==3.0.2
importlib-metadata==7.0.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.7.4
numpy==1.24.4
packaging==23.2
pandas==2.0.3
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.23
typing-extensions==4.9.0
tzdata==2023.3
werkzeug==3.0.1
zipp==3.17.0
